+++ redis/src/lib.rs
@@ -33,28 +33,12 @@ use redis::{
     Client, IntoConnectionInfo, RedisError, RedisResult,
 };
 
-pub use deadpool::managed::reexports::*;
 pub use redis;
 
 pub use self::config::{Config, ConfigError};
 
-/// Type alias for using [`deadpool::managed::Pool`] with [`redis`].
-pub type Pool = managed::Pool<Manager, Connection>;
-
-/// Type alias for using [`deadpool::managed::Pool`] with [`redis`].
-pub type PoolBuilder = managed::PoolBuilder<Manager, Connection>;
-
-/// Type alias for using [`deadpool::managed::PoolError`] with [`redis`].
-pub type PoolError = managed::PoolError<RedisError>;
-
-/// Type alias for using [`deadpool::managed::BuildError`] with [`redis`].
-pub type BuildError = managed::BuildError<RedisError>;
-
-/// Type alias for using [`deadpool::managed::CreatePoolError`] with [`redis`].
-pub type CreatePoolError = managed::CreatePoolError<ConfigError, RedisError>;
-
-/// Type alias for using [`deadpool::managed::Object`] with [`redis`].
-type Object = managed::Object<Manager>;
+pub use deadpool::managed::reexports::*;
+deadpool::managed_reexports!("redis", Manager, Connection, RedisError, ConfigError);
 
 /// Type alias for using [`deadpool::managed::RecycleResult`] with [`redis`].
 type RecycleResult = managed::RecycleResult<RedisError>;