+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Update `tokio` dependency to version `1.5.0`
 * Add `post_create`, `pre_recycle` and `post_recycle` hooks
 * Add `Pool::resize` method
+* Add `managed_reexports` macro
 
 ## v0.8.2